Check descriptions for links
Links can be set by surrounding tehm with pipes. For example |is this
a link|. Some sanity checking is done, there must be an even number of
pipes. At this time links are just highlighted, more is up next.

+++ src/parser.c
@@ -46,8 +46,8 @@ static char
 {
 	char *cur;
 	char *string;
-	int32_t len;
-	int32_t oldlen;
+	size_t len;
+	size_t oldlen;
 
 	assert(tokens);
